fix(abstract-utxo): move toCanonicalTransactionRecipient to parseTransaction
Move toCanonicalTransactionRecipient out of the AbstractUtxoCoin class into
the parseTransaction module where it's used, removing it from the public API.

Issue: BTC-2962
